requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.6
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
